refactor: API 경로 관리
- Interface 추상화를 통해 도메인 별 API 경로 상수처리
