+++ src/main/java/io/hhplus/conbook/config/SecurityConfig.java
@@ -1,12 +1,13 @@
 package io.hhplus.conbook.config;
 
 import io.hhplus.conbook.domain.token.TokenManager;
+import io.hhplus.conbook.interfaces.api.ApiRoutes;
+import io.hhplus.conbook.interfaces.filter.TokenAuthenticationFilter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
@@ -25,20 +26,14 @@ public class SecurityConfig {
                 .httpBasic(basic -> basic.disable())
                 .formLogin(formLogin -> formLogin.disable())
 
-                .securityMatcher("/v1/**")
                 .sessionManagement(session ->
                         session.sessionCreationPolicy(SessionCreationPolicy.STATELESS)
                 )
                 .authorizeHttpRequests(request -> request
-                        .requestMatchers("/v1/**").authenticated()
-                        .requestMatchers("/v1/token/**").permitAll()
+                        .requestMatchers(ApiRoutes.BASE_TOKEN_API_PATH + "/**").permitAll()
+                        .requestMatchers(ApiRoutes.API_V1 + "/**").authenticated()
                 )
                 .addFilterBefore(new TokenAuthenticationFilter(tokenManager), UsernamePasswordAuthenticationFilter.class)
                 .build();
-    }
-
-    @Bean
-    public WebSecurityCustomizer webSecurityCustomizer() {
-        return (web) -> web.ignoring().requestMatchers("/v1/token/**");
     }
 }
