+++ src/main/java/io/hhplus/conbook/interfaces/api/ApiRoutes.java
@@ -1,0 +1,9 @@
+package io.hhplus.conbook.interfaces.api;
+
+public interface ApiRoutes {
+    String API_V1 = "/v1";
+    String BASE_BOOKING_API_PATH = API_V1 + "/bookings";
+    String BASE_CONCERT_API_PATH = API_V1 + "/concerts";
+    String BASE_POINT_API_PATH = API_V1 + "/point";
+    String BASE_TOKEN_API_PATH = API_V1 + "/token";
+}
