+++ src/main/java/io/hhplus/conbook/interfaces/api/booking/BookingController.java
@@ -3,14 +3,15 @@ package io.hhplus.conbook.interfaces.api.booking;
 import io.hhplus.conbook.application.booking.BookingPaymentCommand;
 import io.hhplus.conbook.application.booking.BookingPaymentFacade;
 import io.hhplus.conbook.application.booking.BookingPaymentResult;
-import io.hhplus.conbook.config.CustomAttribute;
+import io.hhplus.conbook.interfaces.api.ApiRoutes;
+import io.hhplus.conbook.interfaces.filter.CustomAttribute;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
 
 @RestController
-@RequestMapping("/v1/bookings")
+@RequestMapping(ApiRoutes.BASE_BOOKING_API_PATH)
 @RequiredArgsConstructor
 public class BookingController implements BookingControllerApi {
